Make bundle for options, split settingsConsole from modules
