+++ lib/modules/menu.js
@@ -6,7 +6,7 @@ import * as Modules from '../core/modules';
 import { i18n } from '../environment';
 import * as Floater from './floater';
 import * as Hover from './hover';
-import * as SettingsConsole from './settingsConsole';
+import * as SettingsNavigation from './settingsNavigation';
 import * as CommandLine from './commandLine';
 
 export const module: Module<*> = new Module('RESMenu');
@@ -66,7 +66,7 @@ function renderConsoleLink() {
 			if (module.options.gearIconClickAction.value === 'openCommandLine' && Modules.isRunning(CommandLine)) {
 				CommandLine.toggle();
 			} else if (module.options.gearIconClickAction.value === 'openSettings') {
-				SettingsConsole.open();
+				SettingsNavigation.loadSettingsPage();
 			} else {
 				showPrefsDropdown(0);
 			}
