+++ lib/modules/search.js
@@ -9,7 +9,6 @@ import { Alert, CreateElement, downcast, empty, frameThrottle, string } from '..
 import { i18n } from '../environment';
 import * as CommandLine from './commandLine';
 import * as Menu from './menu';
-import * as SettingsConsole from './settingsConsole';
 import * as SettingsNavigation from './settingsNavigation';
 
 export const module: Module<*> = new Module('search');
@@ -49,7 +48,7 @@ module.go = () => {
 			input().focus();
 		},
 	});
-	$menuItem.appendTo(SettingsConsole.$menuItem());
+	$menuItem.appendTo(SettingsNavigation.$menuItem());
 
 	CommandLine.registerCommand(/^set(?:t?ings?)?$/, 'settings [words to search for]- search RES settings console',
 		(command, val) => {
@@ -241,7 +240,7 @@ export const input = _.once(() => {
 });
 
 function drawSearchResults(query, results: Array<*>) {
-	SettingsConsole.open(module.moduleID);
+	SettingsNavigation.loadSettingsPage(module.moduleID);
 
 	const $resultsContainer = $('#SearchRES-results-container');
 	$resultsContainer.off('click', handleSearchResultClick)
