+++ lib/modules/settingsNavigation.js
@@ -1,13 +1,13 @@
 /* @flow */
 
+import _ from 'lodash';
 import { $ } from '../vendor';
 import { i18n } from '../environment';
 import { Module } from '../core/module';
 import * as Modules from '../core/modules';
-import * as Options from '../core/options';
-import { Alert } from '../utils';
-import * as Search from './search';
-import * as SettingsConsole from './settingsConsole';
+import * as Menu from '../modules/menu';
+import * as CommandLine from '../modules/commandLine';
+import { string } from '../utils';
 
 export const module: Module<*> = new Module('settingsNavigation');
 
@@ -33,11 +33,31 @@ module.options = {
 	},
 };
 
+export const $menuItem = _.once(() => $('<div>', { id: 'SettingsConsole', text: i18n('RESSettingsConsole') }));
+
+module.beforeLoad = () => {
+	Menu.addMenuItem($menuItem(), () => {
+		Menu.hidePrefsDropdown();
+		loadSettingsPage();
+	}, true);
+};
+
 module.go = () => {
-	window.addEventListener('hashchange', onHashChange);
-	window.addEventListener('popstate', onPopState);
-	// for initial pageload; can be done immediately, because the settings console is ready immediately
-	onHashChange();
+	function findModules(val) {
+		return Modules.all()
+			.filter(v => !v.hidden)
+			.map(v => v.moduleID)
+			.filter(id => id.toLowerCase().match(val.toLowerCase()))
+			.sort();
+	}
+
+	CommandLine.registerCommand(/^mod(?:ule?)?$/, 'module [name of module] - open module settings',
+		(command, val) => {
+			const matches = findModules(val);
+			return matches.length ? `Open module ${matches[0]}` : 'Could not find any matching module.';
+		},
+		(command, val) => open(findModules(val)[0])
+	);
 };
 
 export function makeUrlHashLink(moduleID: string, optionKey?: string, displayText?: string, cssClass?: string): string {
@@ -108,67 +128,17 @@ export function isSettingsUrl(url: string): boolean {
 	return url.includes('#!settings') || url.includes('#res:settings');
 }
 
-function onHashChange() {
-	const hash = location.hash;
-	if (!isSettingsUrl(hash)) return;
-
-	const params = hash.match(/\/(?:\w|\s|%20)+/g);
-	let moduleID;
-	if (params && params[0]) {
-		moduleID = params[0].substring(1).replace('%20', ' ');
-	}
-	let optionKey;
-	if (params && params[1]) {
-		optionKey = params[1].substring(1).replace('%20', ' ');
-	}
-
-	loadSettingsPage(moduleID, optionKey);
-}
-
-function onPopState(event) {
-	const state = typeof event.state === 'string' && event.state.split('/');
-	if (!state || !isSettingsUrl(state[0])) {
-		if (SettingsConsole.isOpen) {
-			// Avoid adding a duplicate page to the browser history
-			SettingsConsole.close({ resetUrl: false });
-		}
-		return;
-	}
-
-	const moduleID = state[1];
-	const optionKey = state[2];
-
-	loadSettingsPage(moduleID, optionKey);
-}
+const iframe = document.createElement('iframe');
 
 export function loadSettingsPage(moduleID?: string, optionKey?: string) {
-	const mod = moduleID && Modules.getUnchecked(moduleID);
-
-	SettingsConsole.open(mod && mod.moduleID);
-
-	if (mod === Search.module) {
-		Search.search(optionKey);
-	} else if (mod) {
-		if (optionKey && mod.options.hasOwnProperty(optionKey)) {
-			const $optionsPanel = $(SettingsConsole.RESConsoleContent);
-			const optionElement = $optionsPanel.find(`label[for="${optionKey}"]`);
-			const optionParent = optionElement.parent();
-			optionParent.addClass('highlight');
-			optionParent.show();
-			if (optionElement.length) {
-				if (optionParent.hasClass('advanced') && !module.options.showAllOptions.value) {
-					document.getElementById('RESConsoleContent').classList.remove('advanced-options-disabled');
-					if (module.options.showAllOptionsAlert.value) {
-						Alert.open('You opened a link to an advanced option, but not all options are shown. These options will be shown until you leave or refresh the page. If you want to see all options in the future, check the <i>Show all options</i> checkbox in the settings console title bar above.<br /><br /><label><input type="checkbox" class="disableAlert" checked="" style="margin:1px 5px 0px 0px;"> Always show this type of notification</label>');
-						$('#alert_message .disableAlert').click(function() {
-							Options.set(module, 'showAllOptionsAlert', this.checked);
-						});
-					}
-				}
-				const $configPanel = $(SettingsConsole.RESConfigPanelOptions);
-				const offset = optionElement.offset().top - $configPanel.offset().top - 10;
-				$configPanel.scrollTop(offset);
-			}
-		}
+	const hash = makeUrlHash(moduleID, optionKey);
+	const url = new URL(hash, chrome.runtime.getURL('options.html'));
+	if (location.origin === url.origin) {
+		location.href = url.href;
+	} else {
+		iframe.hidden = false;
+		iframe.src = url.href;
+		iframe.addEventListener('hashchange', e => console.log(e));
+		document.body.append(iframe);
 	}
 }
