+++ lib/modules/userHighlight.js
@@ -1,6 +1,7 @@
 /* @flow */
 
 import { Module } from '../core/module';
+import * as Options from '../core/options';
 import {
 	Alert,
 	addCSS,
@@ -11,7 +12,6 @@ import {
 	watchForThings,
 } from '../utils';
 import { i18n } from '../environment';
-import * as SettingsConsole from './settingsConsole';
 import * as UserInfo from './userInfo';
 
 export const module: Module<*> = new Module('userHighlight');
@@ -458,12 +458,14 @@ function generateHoverColors() { // apply generateHoverColor on all option
 	const options = ['selfColor', 'OPColor', 'adminColor', 'friendColor', 'modColor', 'firstCommentColor', 'alumColor'];
 	let error = false;
 	for (const option of options) {
-		const newColor = generateHoverColor(SettingsConsole.getOptionValue('userHighlight', option));
-		if (newColor !== false) {
-			SettingsConsole.setOptionValue('userHighlight', `${option}Hover`, newColor);
-		} else {
-			error = true;
-		}
+		// FIXME
+		// const newColor = generateHoverColor(Options.('userHighlight', option));
+		// if (newColor !== false) {
+		// 	void;
+		// 	// SettingsConsole.setOptionValue('userHighlight', `${option}Hover`, newColor);
+		// } else {
+		// 	error = true;
+		// }
 	}
 	if (error) {
 		Alert.open(i18n('userHighlightColorCouldNotBeGenerated'));
