--- lib/options.html
@@ -1,7 +1,0 @@
-<!doctype html>
-<html>
-	<head>
-		<meta http-equiv="refresh" content="0;URL=https://www.reddit.com/#res:settings">
-	</head>
-	<body></body>
-</html>
