+++ lib/options/options.html
@@ -1,0 +1,11 @@
+<!doctype html>
+<html>
+	<head>
+		<meta charset="UTF-8">
+		<title>RES Settings Console</title>
+		<link rel="stylesheet" href="../css/res.scss">
+		<link rel="stylesheet" href="options.scss">
+		<script defer src="options.entry.js"></script>
+	</head>
+	<body></body>
+</html>
