+++ lib/options/settingsConsole.js
@@ -7,7 +7,6 @@ import * as Metadata from '../core/metadata';
 import * as Modules from '../core/modules';
 import * as Options from '../core/options';
 import { $ } from '../vendor';
-import { Module } from '../core/module';
 import {
 	Alert,
 	CreateElement,
@@ -17,20 +16,10 @@ import {
 	niceKeyCode,
 } from '../utils';
 import { i18n } from '../environment';
-import * as About from './about';
-import * as CommandLine from './commandLine';
-import * as Menu from './menu';
-import * as Search from './search';
-import * as SettingsNavigation from './settingsNavigation';
-import { consoleContainerTemplate, moduleSelectorTemplate } from './settingsConsole/templates';
-
-export const module: Module<*> = new Module('settingsConsole');
-
-module.moduleName = 'settingsConsoleName';
-module.category = 'coreCategory';
-module.description = 'settingsConsoleDesc';
-module.alwaysEnabled = true;
-module.hidden = true;
+import * as About from '../modules/about';
+import * as SettingsNavigation from '../modules/settingsNavigation';
+import * as Search from '../modules/search';
+import { consoleContainerTemplate, moduleSelectorTemplate } from './templates';
 
 const DEFAULT_MODULE = About.module;
 const CATEGORY_SORT = [
@@ -53,32 +42,59 @@ export let RESConsoleContent, RESConfigPanelOptions;
 
 export let isOpen = false;
 
-module.beforeLoad = () => {
-	Menu.addMenuItem($menuItem(), () => {
-		Menu.hidePrefsDropdown();
-		open();
-	}, true);
-};
+export function start() {
+	window.addEventListener('hashchange', onHashChange);
+	// for initial pageload; can be done immediately, because the settings console is ready immediately
+	onHashChange();
+}
 
-module.go = () => {
-	function findModules(val) {
-		return Modules.all()
-			.filter(v => !v.hidden)
-			.map(v => v.moduleID)
-			.filter(id => id.toLowerCase().match(val.toLowerCase()))
-			.sort();
+function onHashChange() {
+	const hash = location.hash;
+
+	const params = hash.match(/\/(?:\w|\s|%20)+/g);
+	let moduleID;
+	if (params && params[0]) {
+		moduleID = params[0].substring(1).replace('%20', ' ');
+	}
+	let optionKey;
+	if (params && params[1]) {
+		optionKey = params[1].substring(1).replace('%20', ' ');
 	}
 
-	CommandLine.registerCommand(/^mod(?:ule?)?$/, 'module [name of module] - open module settings',
-		(command, val) => {
-			const matches = findModules(val);
-			return matches.length ? `Open module ${matches[0]}` : 'Could not find any matching module.';
-		},
-		(command, val) => open(findModules(val)[0])
-	);
-};
+	loadSettingsPage(moduleID, optionKey);
+}
 
-export const $menuItem = _.once(() => $('<div>', { id: 'SettingsConsole', text: i18n('RESSettingsConsole') }));
+export function loadSettingsPage(moduleID?: string, optionKey?: string) {
+	const mod = moduleID && Modules.getUnchecked(moduleID);
+
+	open(mod && mod.moduleID);
+
+	if (mod === Search.module) {
+		Search.search(optionKey);
+	} else if (mod) {
+		if (optionKey && mod.options.hasOwnProperty(optionKey)) {
+			const $optionsPanel = $(RESConsoleContent);
+			const optionElement = $optionsPanel.find(`label[for="${optionKey}"]`);
+			const optionParent = optionElement.parent();
+			optionParent.addClass('highlight');
+			optionParent.show();
+			if (optionElement.length) {
+				if (optionParent.hasClass('advanced') && !SettingsNavigation.module.options.showAllOptions.value) {
+					document.getElementById('RESConsoleContent').classList.remove('advanced-options-disabled');
+					if (SettingsNavigation.module.options.showAllOptionsAlert.value) {
+						Alert.open('You opened a link to an advanced option, but not all options are shown. These options will be shown until you leave or refresh the page. If you want to see all options in the future, check the <i>Show all options</i> checkbox in the settings console title bar above.<br /><br /><label><input type="checkbox" class="disableAlert" checked="" style="margin:1px 5px 0px 0px;"> Always show this type of notification</label>');
+						$('#alert_message .disableAlert').click(function() {
+							Options.set(SettingsNavigation.module, 'showAllOptionsAlert', this.checked);
+						});
+					}
+				}
+				const $configPanel = $(RESConfigPanelOptions);
+				const offset = optionElement.offset().top - $configPanel.offset().top - 10;
+				$configPanel.scrollTop(offset);
+			}
+		}
+	}
+}
 
 function create() {
 	// create the console container
@@ -911,7 +927,7 @@ function updateDependsOn(mod) {
 	}
 }
 
-export function open(moduleIdOrCategory?: string) {
+function open(moduleIdOrCategory?: string) {
 	if (!RESConsoleContainer) {
 		create();
 	}
