+++ lib/options/templates.js
@@ -1,7 +1,7 @@
 /* @flow */
 
-import { i18n } from '../../environment';
-import { string } from '../../utils';
+import { i18n } from '../environment';
+import { string } from '../utils';
 
 export const consoleContainerTemplate = ({ name, version, gitDescription }: {| name: string, version: string, gitDescription: string |}) => string.html`
 	<div id="RESConsoleContainer">
